backend/requirements.txt
pymupdf
pymupdf4llm
groq
fastapi
sqlalchemy
pydantic
uvicorn
psycopg2-binary
cloudinary
python-multipart
python-dotenv
alembic